+++ internal/model/order_item.go
@@ -1,9 +1,9 @@
 package model
 
 type OrderItem struct {
-	ID int64 `gorm:"primaryKey" json:"id"`
-	OrderID int64 `json:"orderId"`
-	ProductID int64 `json:"productId"`
+	ID int `gorm:"primaryKey" json:"id"`
+	OrderID int `json:"orderId"`
+	ProductID int `json:"productId"`
 	Product Product `gorm:"foreignKey:ProductID"`
 	Qty int `json:"qty"`
 	UnitPrice float64 `json:"unitPrice"`
